test(h2): split BranchCoverage grab-bag into behavior specs (no prod change)
